fix: address copilot suggestions
