+++ src/utils/clients/tron.ts
@@ -12,5 +12,16 @@ export function createTronClient(rpc: string, serviceId: ServiceID, rpcKey: stri
 }
 
 export async function getLatestBlockNumber(client: TronWeb) {
-  return (await client.trx.getCurrentBlock()).blockID;
+  const block = await client.trx.getCurrentBlock();
+  const blockHeight: unknown = block?.block_header?.raw_data?.number;
+
+  if (typeof blockHeight === "number" && Number.isFinite(blockHeight)) {
+    return blockHeight;
+  }
+
+  if (typeof blockHeight === "string" && blockHeight.trim().length > 0) {
+    return blockHeight;
+  }
+
+  throw new Error("Tron block height is unavailable in current block response");
 }
